requirements.txt
httpx>=0.27.0
requests>=2.32.0
python-dotenv>=1.0.1
pytest>=8.2.0
pytest-asyncio>=0.23.8
respx>=0.21.1

